+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # HEAD
 
 - Added: `at-mixin-argumentless-call-parentheses` rule (with "always"/"never" behavior). Deprecated `at-mixin-no-argumentless-call-parentheses`.
+- Added: `operator-no-unspaced` rule.
 
 # 1.2.1